[it-it] Adapt tests for DateTimeExtractorUtils
